Refactoring.

+++ YasEngineSoftware/CosinusPointsGenerator.cpp
@@ -15,7 +15,8 @@ PointsSet* CosinusPointsGenerator::generatePoints()
 	{
 		float x = i * step;
         float tempY = cos(x);
-		float y = cos(pow(tempY, 2) * 4.5F);
+
+		float y = static_cast<float>( cos(pow(tempY, 2) * 4.5F));
 
         numbers.push_back(tempY);
